-> 05/19/06 - Chris Gilbert (Neovanglist)    * Added gsKit_prim_sprite_striped_texture. This routine autostrips      sprites for a big performance boost. This is most useful on large      sprites which would otherwise result in many GS pagemisses.      NOTE: This ONLY works with textures that have NEAREST filtering.    * Reworked drawbuffers so that it waits on the FINISH signal from the GS      before pushing new DMA to it.    * Added "bigtex" example.    * Added option to delay texture uploading. This allows you to do it manually      using the gsKit_texture_send_inline in usercode.

+++ ee/gs/include/gsFont.h
@@ -47,7 +47,7 @@ int gsKit_font_upload(GSGLOBAL *gsGlobal, GSFONT *gsFont);
 int gsKit_fontm_unpack(GSFONT *gsFont);
 void gsKit_fontm_unpack_raw(unsigned char *base, struct gsKit_fontm_unpack *oke);
 void gsKit_fontm_unpack_raw_1(struct gsKit_fontm_unpack *oke);
-void gsKit_font_print_scaled(GSGLOBAL *gsGlobal, GSFONT *gsFont, float X, float Y, float Z,
+void gsKit_font_print_scaled(GSGLOBAL *gsGlobal, GSFONT *gsFont, float X, float Y, int Z,
 		      float scale, unsigned long color, char *String);
 
 #define gsKit_font_print(gsGlobal, gsFont, X, Y, Z, color, String) \
